refactor: 온보딩 스텝 디렉토리 구조 변경
